requirements.txt
joblib==1.5.2
numpy==2.3.5
nvidia-nccl-cu12==2.28.9
pandas==2.3.3
Pillow==12.0.0
pip==25.2
psutil==7.1.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
xgboost==3.1.2
